app: move init_command_endpoint to app.internal.command_endpoints.main
